動作確認&それに伴う修正
