+++ html/options.html
@@ -35,6 +35,6 @@
         <p><input type="submit" id="btn" value="送信"></p>
     </label>
     <p id="msg"></p>
-    <script src="../dist/option.js"></script>
+    <script src="../dist/options.js"></script>
 </body>
 </html>
