+++ html/popup.html
@@ -6,6 +6,6 @@
 </head>
 <body>
     <button id="btn" class="btn_exe btn--radius">run</button>
-    <script src="../js/popup.js"></script>
+    <script src="../dist/popup.js"></script>
 </body>
 </html>
